Modified according to PR comments and removed test_dataset, since it was renamed to datasim.py
